refector hive files collecting part
make HiveFilesCollector for common use
