+++ src/Storages/Hive/SingleHiveQueryTaskBuilder.h
@@ -1,11 +1,8 @@
 #pragma once
 #include <Common/config.h>
 #if USE_HIVE
-#include <Interpreters/ExpressionActions.h>
 #include <Poco/Logger.h>
-#include <Storages/HDFS/HDFSCommon.h>
 #include <Storages/Hive/HiveQueryTask.h>
-#include <Storages/Hive/StorageHive.h>
 namespace DB
 {
 /**
@@ -23,27 +20,6 @@ public:
 
 private:
     Arguments args;
-    ExpressionActionsPtr partition_key_expr;
-    ExpressionActionsPtr partition_minmax_idx_expr;
-    NamesAndTypesList partition_name_and_types;
-    ExpressionActionsPtr hive_file_minmax_idx_expr;
-    NamesAndTypesList hive_file_name_and_types;
-    String format_name;
-    String hdfs_namenode_url;
-
-    Poco::Logger * logger = &Poco::Logger::get("SingleHiveQueryTaskFilesCollector");
-
-    HiveFiles collectHiveFilesFromPartition(
-        const Apache::Hadoop::Hive::Partition & partition_,
-        SelectQueryInfo & query_info_,
-        HiveMetastoreClient::HiveTableMetadataPtr hive_table_metadata_,
-        const HDFSFSPtr & fs_,
-        ContextPtr context_);
-
-    static ASTPtr extractKeyExpressionList(const ASTPtr & node);
-
-    HiveFilePtr createHiveFileIfNeeded(
-        const HiveMetastoreClient::FileInfo & file_info_, const FieldVector & fields_, SelectQueryInfo & query_info_, ContextPtr context_);
 };
 
 }
